Widen count columns to BIGINT, add search-based creator channel discovery
- db.py: video_stats/channel_stats view/subscriber counts overflowed Postgres
  INTEGER (32-bit) for big channels; switched to BIGINT (also migrated the live
  Supabase columns directly, not just the CREATE TABLE IF NOT EXISTS schema)
- collect_trending.py: channels whose video falls in an excluded category
  (News & Politics, 25) are now auto-excluded instead of auto-included; the
  video/stats/rank data is still collected either way
- youtube_client.py: add search_list (search.list, order support)
- discover_channels.py: new one-off script — searches genre keywords with
  order=viewCount, filters out excluded categories, and seeds the channel pool
  with high-view creator channels instead of relying only on what surfaces in
  daily trending. Run against real Supabase: 30 keywords, 852 unique channels,
  3030 quota units
- settings.toml: [discovery] keyword list + quota costs,
  [trending].exclude_channel_categories
- also retroactively excluded 61 pre-existing news-category channels and
  paused the GitHub Actions schedule (channel curation strategy was being
  reworked) — see docs/collect-channels-spec.md for context

+++ src/db.py
@@ -47,9 +47,9 @@ CREATE TABLE IF NOT EXISTS video_stats (
   video_id      TEXT NOT NULL,
   collected_at  TEXT NOT NULL,
   run_id        TEXT NOT NULL,
-  view_count    INTEGER,
-  like_count    INTEGER,
-  comment_count INTEGER,
+  view_count    BIGINT,
+  like_count    BIGINT,
+  comment_count BIGINT,
   PRIMARY KEY (video_id, collected_at)
 );
 
@@ -80,8 +80,8 @@ CREATE TABLE IF NOT EXISTS channel_stats (
   channel_id       TEXT NOT NULL,
   collected_at     TEXT NOT NULL,
   run_id           TEXT NOT NULL,
-  subscriber_count INTEGER,
-  view_count       INTEGER,
+  subscriber_count BIGINT,
+  view_count       BIGINT,
   video_count      INTEGER,
   PRIMARY KEY (channel_id, collected_at)
 );
